h264nal: fix sign conversion warnings
Tested:
```
$ mkdir build; cd build; CC=clang CXX=clang++  cmake -DCMAKE_BUILD_TYPE=DEBUG ..
$ make -j
...
[100%] Built target h264nal-bin

$ make test
Running tests...
Test project h264nal/build
      Start  1: h264_common_unittest
 1/23 Test  #1: h264_common_unittest .........................................   Passed    0.00 sec
      Start  2: h264_hrd_parameters_parser_unittest
 2/23 Test  #2: h264_hrd_parameters_parser_unittest ..........................   Passed    0.00 sec
      Start  3: h264_vui_parameters_parser_unittest
 3/23 Test  #3: h264_vui_parameters_parser_unittest ..........................   Passed    0.00 sec
      Start  4: h264_sps_parser_unittest
 4/23 Test  #4: h264_sps_parser_unittest .....................................   Passed    0.00 sec
      Start  5: h264_pps_parser_unittest
 5/23 Test  #5: h264_pps_parser_unittest .....................................   Passed    0.00 sec
      Start  6: h264_ref_pic_list_modification_parser_unittest
 6/23 Test  #6: h264_ref_pic_list_modification_parser_unittest ...............   Passed    0.00 sec
      Start  7: h264_pred_weight_table_parser_unittest
 7/23 Test  #7: h264_pred_weight_table_parser_unittest .......................   Passed    0.00 sec
      Start  8: h264_dec_ref_pic_marking_parser_unittest
 8/23 Test  #8: h264_dec_ref_pic_marking_parser_unittest .....................   Passed    0.00 sec
      Start  9: h264_rtp_single_parser_unittest
 9/23 Test  #9: h264_rtp_single_parser_unittest ..............................   Passed    0.00 sec
      Start 10: h264_rtp_stapa_parser_unittest
10/23 Test #10: h264_rtp_stapa_parser_unittest ...............................   Passed    0.00 sec
      Start 11: h264_rtp_fua_parser_unittest
11/23 Test #11: h264_rtp_fua_parser_unittest .................................   Passed    0.00 sec
      Start 12: h264_rtp_parser_unittest
12/23 Test #12: h264_rtp_parser_unittest .....................................   Passed    0.00 sec
      Start 13: h264_slice_header_parser_unittest
13/23 Test #13: h264_slice_header_parser_unittest ............................   Passed    0.00 sec
      Start 14: h264_slice_header_in_scalable_extension_parser_unittest
14/23 Test #14: h264_slice_header_in_scalable_extension_parser_unittest ......   Passed    0.00 sec
      Start 15: h264_slice_layer_without_partitioning_rbsp_parser_unittest
15/23 Test #15: h264_slice_layer_without_partitioning_rbsp_parser_unittest ...   Passed    0.00 sec
      Start 16: h264_slice_layer_extension_rbsp_parser_unittest
16/23 Test #16: h264_slice_layer_extension_rbsp_parser_unittest ..............   Passed    0.00 sec
      Start 17: h264_sps_extension_parser_unittest
17/23 Test #17: h264_sps_extension_parser_unittest ...........................   Passed    0.00 sec
      Start 18: h264_subset_sps_parser_unittest
18/23 Test #18: h264_subset_sps_parser_unittest ..............................   Passed    0.00 sec
      Start 19: h264_sps_svc_extension_parser_unittest
19/23 Test #19: h264_sps_svc_extension_parser_unittest .......................   Passed    0.00 sec
      Start 20: h264_bitstream_parser_unittest
20/23 Test #20: h264_bitstream_parser_unittest ...............................   Passed    0.00 sec
      Start 21: h264_prefix_nal_unit_parser_unittest
21/23 Test #21: h264_prefix_nal_unit_parser_unittest .........................   Passed    0.00 sec
      Start 22: h264_nal_unit_header_svc_extension_parser_unittest
22/23 Test #22: h264_nal_unit_header_svc_extension_parser_unittest ...........   Passed    0.00 sec
      Start 23: h264_nal_unit_parser_unittest
23/23 Test #23: h264_nal_unit_parser_unittest ................................   Passed    0.00 sec

100% tests passed, 0 tests failed out of 23

Total Test time (real) =   0.07 sec
```

+++ src/h264_common.cc
@@ -257,7 +257,7 @@ bool byte_aligned(BitBuffer *bit_buffer) {
   return (out_bit_offset == 0);
 }
 
-int get_current_offset(BitBuffer *bit_buffer) {
+size_t get_current_offset(BitBuffer *bit_buffer) {
   size_t out_byte_offset, out_bit_offset;
   bit_buffer->GetCurrentOffset(&out_byte_offset, &out_bit_offset);
 
@@ -389,7 +389,8 @@ std::shared_ptr<NaluChecksum> NaluChecksum::GetNaluChecksum(
   val = 0;
   while (bit_buffer->RemainingBitCount() > 0) {
     (void)bit_buffer->ReadUInt8(val8);
-    val |= (val8 << (8 * (3 - i)));
+    val |= static_cast<uint32_t>(val8)
+           << (8 * static_cast<unsigned int>(3 - i));
     i += 1;
   }
   if (i > 0) {
@@ -423,7 +424,7 @@ void NaluChecksum::fdump(char *output, int output_len) const {
       output[output_len - 1] = '\0';
       break;
     }
-    oi += snprintf(output + oi, output_len - oi, "%02x",
+    oi += snprintf(output + oi, static_cast<size_t>(output_len - oi), "%02x",
                    static_cast<unsigned char>(checksum[i++]));
   }
 }
@@ -434,7 +435,7 @@ const char *NaluChecksum::GetPrintableChecksum() const {
   int i = 0;
   int oi = 0;
   while (i < length) {
-    oi += snprintf(buffer + oi, BUFFER_LEN - oi, "%02x",
+    oi += snprintf(buffer + oi, static_cast<size_t>(BUFFER_LEN - oi), "%02x",
                    static_cast<unsigned char>(checksum[i++]));
   }
   buffer[i] = '\0';
