+++ test/h264_bitstream_parser_unittest.cc
@@ -59,7 +59,7 @@ TEST_F(H264BitstreamParserTest, TestSampleBitstream601) {
   EXPECT_EQ(4, bitstream->nal_units.size());
 
   // check the 1st NAL unit
-  int index = 0;
+  size_t index = 0;
   EXPECT_EQ(0,
             bitstream->nal_units[index]->nal_unit_header->forbidden_zero_bit);
   EXPECT_EQ(3, bitstream->nal_units[index]->nal_unit_header->nal_ref_idc);
@@ -125,7 +125,7 @@ TEST_F(H264BitstreamParserTest, TestSampleBitstream601Alt) {
   int length = 0;
   int counter = 0;
   // 1st NAL unit
-  int index = 0;
+  size_t index = 0;
   length = 24;
   counter += 4;
   EXPECT_EQ(counter, bitstream->nal_units[index]->offset);
